Fix project structure and add .gitignore
