--- Real-time-Node-Site/RTNS/Views/footer.ejs
@@ -1,2 +1,0 @@
-<div>V 1.0 - 12/22/2017 - No Rights Reserved</div>
-<div><a href="/admin">Admin Login</a></div>
