--- Real-time-Node-Site/RTNS/package.json
@@ -1,11 +1,0 @@
-{
-  "private": true,
-  "dependencies": {
-    "body-parser": "^1.18.2",
-    "ejs": "^2.5.7",
-    "express": "^4.16.2",
-    "express-session": "^1.15.6",
-    "mongodb": "^3.0.0-rc0",
-    "socket.io": "^2.0.4"
-  }
-}
