--- Real-time-Node-Site/Views/header.ejs
@@ -1,13 +1,0 @@
-<div><h2><a href="/">Node Template Site</a></h2></div>
-<!--- If user is logged in--->
-<% if (session.username) { %>
-  <div>Welcome, <%= session.username %>. | <a href='/sign-out'>Sign Out</a></div>
-<% }else{ %>
-  <div id="login-div">
-    <form method="POST" action="/log-in">
-      <input name="username" placeholder="username" type="text"/>
-      <input name="password" placeholder="password" type="password"/>
-      <input value="Login" type="submit"> or <a href='/register'>Register</a>
-    </form>
-  </div>
-<% } %>
