--- Real-time-Node-Site/public/client.js
@@ -1,16 +1,0 @@
-console.log('Hello');
-const socket = io();
-
-// Ask server for users data
-socket.emit('want-users-list');
-
-// Listen for users array replacement from server.
-socket.on('users-list', function (usersArray) {
-  var listElement = document.getElementById('usersList');
-  // For each user in Users array...
-  usersArray.forEach((user) => {
-    const newElement = document.createElement('li');
-    newElement.textContent = user.username;
-    listElement.appendChild(newElement);
-  });
-});
