+++ Views/footer.ejs
@@ -1,0 +1,2 @@
+<div>V 1.0 - 12/22/2017 - No Rights Reserved</div>
+<div><a href="/admin">Admin Login</a></div>
