+++ Views/index.ejs
@@ -1,0 +1,10 @@
+<%- include('header'); %>
+<body>
+  <h3>A Humble Introduction</h3>
+  <p>
+    Welcome to this minimal Node Template Site, which strives to present a<br/>
+    simple to use real time template site which new developers may use as a<br/>
+    foundation to build new Node.JS applications upon.
+  </p>
+</body>
+<%- include('footer'); %>
